Create filters.unset() allowing the user to remove filters
Previously, the only way of doing so was to use ```.set()``` removing whatever
you wanted. If you wanted to remove entirely the filter, you would
```.set('filter_name', '')```, but it would still stay on the URL as
```?filters=filter_name:```

With ```unset``` we properly remove the filter.

In the future, it would be good to detect if we removed all filters, and remove
the ```filters``` key itself from the URL, so we don't end up with something
like ```?filters=```.